requirements.txt
Flask==2.1.0
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
PyMySQL==1.0.2
python-dotenv==0.19.2
requests==2.26.0
